1. Fixed #284 by changing data type of MouseEvent.x, MouseEvent.y, MouseEvent.clientX, MouseEvent.clientY to float. 2. Added an exception handler to kwebConfig that allows to react to errors during the processing of websocket errors. 3. removed the throw exception part in debugErr in kweb_bootstrap as is intentionally breaks the client in case of an exception which is completely unnecessary if the error is not servere.

+++ src/main/kotlin/kweb/html/events/events.kt
@@ -44,8 +44,8 @@ data class MouseEvent(
     val altKey: Boolean,
     val button: Int,
     val buttons: Int,
-    val clientX: Int,
-    val clientY: Int,
+    val clientX: Float,
+    val clientY: Float,
     val ctrlKey: Boolean,
     val metaKey: Boolean,
     val movementX: Int? = null,
@@ -54,8 +54,8 @@ data class MouseEvent(
     val screenX: Int,
     val screenY: Int,
     val shiftKey: Boolean,
-    val x: Int = clientX,
-    val y: Int = clientY,
+    val x: Float = clientX,
+    val y: Float = clientY,
     /** @see kweb.Element.on **/
     val retrieved: JsonElement = JsonNull
 )
